Move listEncryptedColumns method to a function.
